replaced all int with int16_t
to avoid problem on 32bit microcontroler platforms

+++ TICC/board.h
@@ -12,36 +12,36 @@
 
 // I/0 assignments
 #if (BOARD_REVISION == 'C')  // rev C board had different mapping for these pins
-const int OUT1 =        2;   // spare output
-const int OUT2 =        3;   // spare output
-const int STOP_0 =      12;
-const int STOP_1 =      13;
+const int16_t OUT1 =        2;   // spare output
+const int16_t OUT2 =        3;   // spare output
+const int16_t STOP_0 =      12;
+const int16_t STOP_1 =      13;
 #elif (BOARD_REVISION == 'D')
-const int STOP_0 =      2;   // PINE,4
-const int STOP_1 =      3;   // PINE,5
-const int OUT1 =        12;  // spare output
-const int OUT2 =        13;  // spare output
+const int16_t STOP_0 =      2;   // PINE,4
+const int16_t STOP_1 =      3;   // PINE,5
+const int16_t OUT1 =        12;  // spare output
+const int16_t OUT2 =        13;  // spare output
 #endif
 
-const int ENABLE_0 = 	  4;
-const int ENABLE_1 =    5;
-const int CSB_0  =      6;
-const int CSB_1 =       7;
-const int INP1 =        8;   // spare input
-const int INP2 =        9;   // spare input
-const int INTB_0 =  	  10;  // PINB,4
-const int INTB_1 =      11;  // PINB,5
-const int D16 =         16;  // spare unassigned
-const int D17 =         17;  // spare unassigned
-const int COARSEint =   18;  // hardware interrupt for COARSE clock
-const int CLIENT_SYNC =  A8;  // use to sync multiple boards
-const int AN9 =         A9;  // spare unassigned
-const int AN10 =        A10; // spare unassigned
-const int EXT_LED_CLK = A11; // external LED shows 100kHz clock status -- PORTK,3
-const int EXT_LED_0 =   A12; // external LED tandem with LED_0 -- PORTK,4
-const int EXT_LED_1 =   A13; // external LED tandem with LED_1 -- PORTK,5
-const int LED_0 =       A14; // onboard LED -- PORTK,6
-const int LED_1 =       A15; // onboard LED -- PORTK,7
+const int16_t ENABLE_0 = 	  4;
+const int16_t ENABLE_1 =    5;
+const int16_t CSB_0  =      6;
+const int16_t CSB_1 =       7;
+const int16_t INP1 =        8;   // spare input
+const int16_t INP2 =        9;   // spare input
+const int16_t INTB_0 =  	  10;  // PINB,4
+const int16_t INTB_1 =      11;  // PINB,5
+const int16_t D16 =         16;  // spare unassigned
+const int16_t D17 =         17;  // spare unassigned
+const int16_t COARSEint =   18;  // hardware interrupt for COARSE clock
+const int16_t CLIENT_SYNC =  A8;  // use to sync multiple boards
+const int16_t AN9 =         A9;  // spare unassigned
+const int16_t AN10 =        A10; // spare unassigned
+const int16_t EXT_LED_CLK = A11; // external LED shows 100kHz clock status -- PORTK,3
+const int16_t EXT_LED_0 =   A12; // external LED tandem with LED_0 -- PORTK,4
+const int16_t EXT_LED_1 =   A13; // external LED tandem with LED_1 -- PORTK,5
+const int16_t LED_0 =       A14; // onboard LED -- PORTK,6
+const int16_t LED_1 =       A15; // onboard LED -- PORTK,7
 
 // These are macros to turn LEDs on and off really fast.
 // We trade flexibility for speed.
